# specs/002-attendance-account-shift-leave/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 出勤加班單系統 API（擴充需求 3／線 B）
  description: |
    前後端對齊契約。瀏覽器經 Nginx 同源路徑為 `/attendance/api/...`；
    本機直接打 Fastify 時 server 為 `http://localhost:3000`，path 仍為 `/api/...`。

    Cookie 名稱：`attendance_session`（HttpOnly）。登入成功後後續請求自動帶入。

    **登入驗證（2026-09-20｜對齊教育訓練）**：`GET /api/auth/captcha` 取圖形驗證碼；
    員工＝編號＋驗證碼；Admin＝帳號＋密碼＋驗證碼。**不使用入庫 PIN。**

    **Auth 範圍（PRD §5.3.1｜2026-09-20）**：加班單主流程（上傳／計算／預覽／匯出）
    為前端公開能力，**不**經本 API 強制登入。本契約端點（行事曆、Admin、
    伺服器正式匯入、詞庫寫入等）仍需 session；`POST /api/attendance/import`
    為寫入伺服器台帳之用，與「未登入本機公開上傳」不同層。

    錯誤回應一律為 `ErrorResponse`。常用 `code`：
    `UNAUTHORIZED`、`FORBIDDEN`、`VALIDATION_ERROR`、`NOT_FOUND`、
    `LOGIN_LOCKED`、`IMPORT_NOT_SELF`、`UNKNOWN_LEAVE_TYPE`、
    `SHIFT_IN_USE`、`SHIFT_HAS_ACTIVE_ASSIGNMENT`、`ADMIN_PROTECTED`。
  version: 0.1.0
  contact:
    name: 出勤加班單系統
servers:
  - url: http://localhost:3000
    description: 本機 Fastify（開發）
  - url: /attendance
    description: 經 Nginx 反代（與前端同源）
...
paths:
...
  /api/admin/employees:
    get:
      tags: [admin-employees]
      summary: 員工列表（不含 Admin）
      description: 回傳一般員工帳號；role=admin（含固定超管 000000）不列入。
      operationId: listEmployees
      x-phase: B1
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/QueryActiveOnly'
      responses:
        '200':
          description: OK（不含 Admin）
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin-employees]
      summary: 新增員工
      operationId: createEmployee
      x-phase: B1
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/employees/{employeeId}:
    parameters:
      - $ref: '#/components/parameters/EmployeeIdPath'
    get:
      tags: [admin-employees]
      summary: 員工詳情（含年假額度）
      operationId: getEmployee
      x-phase: B1
      security:
        - sessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [admin-employees]
      summary: 更新員工（啟停、年假額度、Admin 密碼）
      operationId: updateEmployee
      x-phase: B1
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 非 Admin，或試圖停用／刪除受保護 Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/shifts:
    get:
      tags: [admin-shifts]
      summary: 班表列表
      operationId: listShifts
      x-phase: B2
      security:
        - sessionCookie: []
      parameters:
        - name: includeDisabled
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Shift'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin-shifts]
      summary: 新增班表
      operationId: createShift
      x-phase: B2
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftWriteRequest'
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/shifts/{shiftId}:
    parameters:
      - $ref: '#/components/parameters/ShiftIdPath'
    patch:
      tags: [admin-shifts]
      summary: 更新／停用班表
      operationId: updateShift
      x-phase: B2
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 現職仍有派班，無法停用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                activeAssignment:
                  value:
                    code: SHIFT_HAS_ACTIVE_ASSIGNMENT
                    message: 目前仍有員工派在此班，無法停用
    delete:
      tags: [admin-shifts]
      summary: 刪除班表（僅無任何歷史引用時）
      operationId: deleteShift
      x-phase: B2
      security:
        - sessionCookie: []
      responses:
        '204':
          description: 已刪除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 仍有派班或歷史計算引用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                inUse:
                  value:
                    code: SHIFT_IN_USE
                    message: 班表已有引用，不可刪除，僅可停用
  /api/admin/shift-assignments:
    get:
      tags: [admin-shifts]
      summary: 派班列表（不含 Admin）
      description: 回傳一般員工派班列；Admin／超管使用者的派班不列入。
      operationId: listShiftAssignments
      x-phase: B2
      security:
        - sessionCookie: []
      parameters:
        - name: employeeId
          in: query
          schema:
            $ref: '#/components/schemas/EmployeeId'
      responses:
        '200':
          description: OK（不含 Admin）
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ShiftAssignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin-shifts]
      summary: 新增派班起迄
      description: 不可對 role=admin 或超管員工編號（000000）派班；違反時回 400。
      operationId: createShiftAssignment
      x-phase: B2
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftAssignmentWriteRequest'
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftAssignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/shift-assignments/{assignmentId}:
    parameters:
      - $ref: '#/components/parameters/AssignmentIdPath'
    patch:
      tags: [admin-shifts]
      summary: 更新派班起迄
      description: 不可將派班改指到 role=admin 或超管員工編號（000000）；違反時回 400。
      operationId: updateShiftAssignment
      x-phase: B2
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftAssignmentWriteRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftAssignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin-shifts]
      summary: 刪除派班區段
      operationId: deleteShiftAssignment
      x-phase: B2
      security:
        - sessionCookie: []
      responses:
        '204':
          description: 已刪除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/work-locations:
    get:
      tags: [work-locations]
      summary: 工作地點詞庫前綴搜尋
      operationId: searchWorkLocations
      x-phase: B5
      security:
        - sessionCookie: []
      parameters:
        - name: q
          in: query
          required: true
          description: 前綴；可為單一字元
          schema:
            type: string
            minLength: 1
            maxLength: 40
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkLocationTerm'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [work-locations]
      summary: 新增工作地點字串至共用詞庫
      description: 已存在則回傳既有項（冪等），不報錯。
      operationId: createWorkLocation
      x-phase: B5
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 40
      responses:
        '200':
          description: 已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLocationTerm'
        '201':
          description: 新建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLocationTerm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: attendance_session
  parameters:
    EmployeeIdPath:
      name: employeeId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/EmployeeId'
    ShiftIdPath:
      name: shiftId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssignmentIdPath:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryActiveOnly:
      name: activeOnly
      in: query
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: 請求驗證失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未登入或 session 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: UNAUTHORIZED
                message: 請先登入
    Forbidden:
      description: 已登入但無權限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                code: FORBIDDEN
                message: 權限不足
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: NOT_FOUND
                message: 找不到資源
  schemas:
    EmployeeId:
      type: string
      pattern: '^\d{6}$'
      example: '100057'
    IsoDate:
      type: string
      format: date
      example: '2026-03-25'
    TimeHm:
      type: string
      pattern: '^\d{2}:\d{2}$'
      example: '09:00'
    Role:
      type: string
      enum: [employee, admin]
    ShiftStatus:
      type: string
      enum: [active, disabled]
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    AnnualLeaveQuota:
      type: object
      required: [year, quotaDays, usedDays, remainingDays]
      properties:
        year:
          type: integer
          example: 2026
        quotaDays:
          type: number
          format: double
        usedDays:
          type: number
          format: double
          description: 該曆年「請年休假」考勤數量加總
        remainingDays:
          type: number
          format: double
          description: quotaDays − usedDays，可為負
    Employee:
      type: object
      required: [employeeId, name, role, isActive, quotaYear]
      properties:
        employeeId:
          $ref: '#/components/schemas/EmployeeId'
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        isActive:
          type: boolean
        quotaYear:
          type: integer
          description: 列表／詳情帶出的當前年（西元）
        quotaDays:
          type: number
          format: double
          nullable: true
          description: 當前年年假額度天數；無紀錄時為 null
    EmployeeDetail:
      allOf:
        - $ref: '#/components/schemas/Employee'
        - type: object
          properties:
            quotas:
              type: array
              items:
                $ref: '#/components/schemas/AnnualLeaveQuota'
    EmployeeCreateRequest:
      type: object
      required: [employeeId, name]
      properties:
        employeeId:
          $ref: '#/components/schemas/EmployeeId'
        name:
          type: string
          minLength: 1
        isActive:
          type: boolean
          default: true
        quotaYear:
          type: integer
          description: 建立時一併寫入的曆年；省略則用伺服器當前年
        quotaDays:
          type: number
          format: double
          default: 0
    EmployeeUpdateRequest:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        password:
          type: string
          description: 僅 Admin 帳號可設／改密碼
        quota:
          type: object
          required: [year, quotaDays]
          properties:
            year:
              type: integer
            quotaDays:
              type: number
              format: double
    Shift:
      type: object
      required: [id, name, startTime, endTime, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 公司班
        startTime:
          $ref: '#/components/schemas/TimeHm'
        endTime:
          $ref: '#/components/schemas/TimeHm'
          description: 平日晚段加班起點
        status:
          $ref: '#/components/schemas/ShiftStatus'
    ShiftWriteRequest:
      type: object
      required: [name, startTime, endTime]
      properties:
        name:
          type: string
        startTime:
          $ref: '#/components/schemas/TimeHm'
        endTime:
          $ref: '#/components/schemas/TimeHm'
    ShiftUpdateRequest:
      type: object
      properties:
        name:
          type: string
        startTime:
          $ref: '#/components/schemas/TimeHm'
        endTime:
          $ref: '#/components/schemas/TimeHm'
        status:
          $ref: '#/components/schemas/ShiftStatus'
    ShiftAssignment:
      type: object
      required: [id, employeeId, employeeName, shiftId, effectiveFrom]
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          $ref: '#/components/schemas/EmployeeId'
        employeeName:
          type: string
          description: 員工顯示姓名
        shiftId:
          type: string
          format: uuid
        shiftName:
          type: string
        effectiveFrom:
          $ref: '#/components/schemas/IsoDate'
        effectiveTo:
          allOf:
            - $ref: '#/components/schemas/IsoDate'
          nullable: true
          description: null 表示尚無結束日
    ShiftAssignmentWriteRequest:
      type: object
      required: [employeeId, shiftId, effectiveFrom]
      properties:
        employeeId:
          $ref: '#/components/schemas/EmployeeId'
        shiftId:
          type: string
          format: uuid
        effectiveFrom:
          $ref: '#/components/schemas/IsoDate'
        effectiveTo:
          allOf:
            - $ref: '#/components/schemas/IsoDate'
          nullable: true
...
    WorkLocationTerm:
      type: object
      required: [id, text, createdAt]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
          maxLength: 40
        createdAt:
          type: string
          format: date-time